+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-[Full diff](https://github.com/sider/runners/compare/0.7.4...HEAD)
+[Full diff](https://github.com/sider/runners/compare/0.7.5...HEAD)
+
+## 0.7.5
+
+[Full diff](https://github.com/sider/runners/compare/0.7.4...0.7.5)
 
 - Improve deprecation message for older versions [#419](https://github.com/sider/runners/pull/419)
 - [SwiftLint] Treat as success when no linting files [#422](https://github.com/sider/runners/pull/422)